revert changes made by https://github.com/lucee/Lucee/commit/1582c4f1663b3a8d35681fe9de7115bd4de78550 for multiple reasons - they are only partial, the loader is updated to jakarta 11, but the core still points to 10 what breaks the build process - there is no backward compatibility, means after this change you need at least jakarta 11 at runtime to run Lucee, it is not necessary to remove methods like `PageExceptionImpl.getRootCause()` even we update the build to jakarta 11, simply remove `@Override` and mark them - Lucee 7.0 is a release candidate, it is to late in the cycle for changes like this, we can do that with Lucee 7.0+
we can make the changes in a way you then still can use jakarta 10 at runtime, but for the moment i prefer to keep it on 10 for the build process but add all the necessary parts so it can also run on 11 at runtime.
i kept `HttpServletResponseDummy.sendRedirect(String location, int status, boolean clearBuffer)` for example is it still is comaptible with jakarta 11.